requirements.txt
aiogram==3.13.1
redis==5.0.8
asyncpg==0.29.0
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.24.0
fakeredis==2.24.0